requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.131.0
h11==0.16.0
idna==3.11
kaggle==2.0.0
kagglesdk==0.1.15
numpy==2.4.2
packaging==26.0
pandas==3.0.1
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
requests==2.32.5
six==1.17.0
starlette==0.52.1
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
webencodings==0.5.1
